refactor: 최종 수강료 계산 및 안내 메일 전송 배치 구현 - second commit
- 이메일 상수로 선언
- batch 패키지 구조 세분화
- 트랜잭션 처리
- 의존성 자동 주입
- 로깅
