+++ src/main/java/com/monari/monariback/enrollment/batch/itemwriter/EnrollmentItemWriter.java
@@ -1,13 +1,17 @@
-package com.monari.monariback.enrollment.batch;
+package com.monari.monariback.enrollment.batch.itemwriter;
 
+import com.monari.monariback.enrollment.dto.LessonFeeDto;
 import com.monari.monariback.enrollment.entity.Enrollment;
 import com.monari.monariback.enrollment.repository.EnrollmentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.batch.item.Chunk;
 import org.springframework.batch.item.ItemWriter;
+import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
+@Component
 @RequiredArgsConstructor
 public class EnrollmentItemWriter implements ItemWriter<LessonFeeDto> {
 
@@ -18,6 +22,7 @@ public class EnrollmentItemWriter implements ItemWriter<LessonFeeDto> {
      * @param chunk of LessonFeeDto - 쓰기 작업의 대상이 될 수업들
      */
     @Override
+    @Transactional
     public void write(Chunk<? extends LessonFeeDto> chunk) throws Exception {
         for (LessonFeeDto dto : chunk) {
             List<Enrollment> enrollments = enrollmentRepository.findAllByLessonId(dto.lessonId());
