+++ src/main/java/com/monari/monariback/enrollment/constant/EnrollmentBatchConstants.java
@@ -1,0 +1,6 @@
+package com.monari.monariback.enrollment.constant;
+
+public class EnrollmentBatchConstants {
+    public static final String SERVICE_EMAIL_ADDRESS = "monari@gmail.com";
+
+}
